remove unnecessary inputs
